Add missing project READMEs and link all project docs from root README
